ConnectionDeleteEditPolicy moved (aurusov/raox#334)
