+++ ru.bmstu.rk9.rao.ui/src/ru/bmstu/rk9/rao/ui/process/node/ConnectionDeleteEditPolicy.java
@@ -1,4 +1,4 @@
-package ru.bmstu.rk9.rao.ui.process;
+package ru.bmstu.rk9.rao.ui.process.node;
 
 import org.eclipse.gef.commands.Command;
 import org.eclipse.gef.editpolicies.ConnectionEditPolicy;
